More sized type cleanup

+++ src/common/IvGraphics/D3D11/IvConstantTableD3D11.h
@@ -43,7 +43,7 @@ struct IvConstantDesc
         };
         struct
         {
-            int           mTextureSlot;
+            intptr_t      mTextureSlot;
             int           mSamplerSlot;
         };
     };
